Extract event details into its own component
